message's classes are now Mongo Documents
